v0.1 - Hello world! Initial commit, containing many initial improvements.
This first commit contains the Blender file and the textures that come with it. So far, compared to the remaster made by Ikeiwa, the only major difference is that the model is 99% symmetrical now, both for the vertices and their weights. I also fixed some weird issues with the left elbow and the right ear. Much more to come!

Many things have been left out or straight up removed, so this first commit is just a way of getting what I've already done out there.